# bmOverwrite fast paths, move some stuff