--- a/experiments/svg_cairo.nim
+++ b/experiments/svg_cairo.nim
@@ -1,6 +1,7 @@
 ## Load and Save SVG files.
 
-import cairo, chroma, pixie/common, pixie/images, pixie/paints, strutils, tables, vmath, xmlparser, xmltree
+import cairo, chroma, pixie/common, pixie/images, pixie/paints, strutils,
+    tables, vmath, xmlparser, xmltree
 
 include pixie/paths