… tests/paths/pathHeart2.png
… tests/test_paths.nim